Add support for grubenv in /boot/writable
On read-only system on error message appeared when writing the grubenv file
(tough this didn't affect functionality); a general mechanism to allow writing
to grubenv was introduced now to avoid this (and allows other applications to
also store their settings).
The main part of the implementation was done in read-only-root-fs,
health-checker just has to mount the subvolume during initrd.

[boo#1126900]