Save last working snapshot number into grubenv
Since the introduction of a unified /var partition in MicroOS / Kubic, the
detection of a broken kernel or initrd did not work any more, as the GRUB
part of health-checker expected the /var partition to be a subvolume of the
root partition.

Instead of trying to somehow find out the correct partition simply save the
latest successfully booted snapshot number into the grub environment block.
The value only has to be read (and not written) from within GRUB.

In case an upgrade from an older release, not using the grubenv variable
yet, is performed (e.g. when updating a transactional server from Leap 15.1
to 15.2) the file has to be read from the old location, so the old code will
stay in there for now.